Refactor libft

+++ libft/ft_atol.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_atoi.c                                          :+:      :+:    :+:   */
+/*   ft_atol.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: afaddoul <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: ybahlaou <ybahlaou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/16 12:26:02 by afaddoul          #+#    #+#             */
-/*   Updated: 2019/09/27 15:00:23 by afaddoul         ###   ########.fr       */
+/*   Updated: 2019/10/02 11:21:32 by ybahlaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long long		ft_atol(const char *str)
+long	ft_atol(const char *str)
 {
 	int			i;
 	int			sign;
